+++ src/auth/client-authentication.ts
