+++ src/auth/index.ts
@@ -1,13 +1,13 @@
-import { TelemetryMiddleware } from '../lib/middleware/TelemetryMiddleware';
-import { Options } from './BaseAuthApi';
-import { Database } from './Database';
-import { OAuth } from './OAuth';
-import { Passwordless } from './Passwordless';
+import { TelemetryMiddleware } from '../lib/middleware/telemetry-middleware';
+import { Options } from './base-auth-api';
+import { Database } from './database-';
+import { OAuth } from './oauth-';
+import { Passwordless } from './passwordless-';
 
-export * from './Database';
-export * from './OAuth';
-export * from './Passwordless';
-export { AuthApiError } from './BaseAuthApi';
+export * from './database-';
+export * from './oauth-';
+export * from './passwordless-';
+export { AuthApiError } from './base-auth-api';
 
 export class AuthenticationClient {
   database: Database;
