+++ src/lib/middleware/telemetry-middleware.ts
